+++ CHANGELOG.md
@@ -9,11 +9,14 @@ All notable changes qsf-integration project will be documented in this file.
 - implement a include for the json builder
 - jsonBuilder: implement an exists function
 - jsonBuilder: implement a function to create not existing paths pathsForceCreate
+- elastic QsfqlFilterTransformer - make the location of the filters configurable with filterPath and filterVariable
 
 ### Removed
 
 ### Changed
 
+### Breaking Changes
+-  the filters can not be set automatically in a combination with a function_score query. For this purpose a filterPath or filterVariable must be configured
 ### Deprecated
 
 ### Fixed